implemented generic views
