+++ src/view/cx.rs
@@ -1,0 +1,89 @@
+use std::collections::HashSet;
+use std::sync::Arc;
+use std::sync::mpsc::SyncSender;
+use std::task::Waker;
+use futures_task::ArcWake;
+use crate::Id;
+use crate::id::IdPath;
+
+#[derive(Clone)]
+pub struct Cx {
+    id_path: IdPath,
+    req_chan: SyncSender<IdPath>,
+    pub(crate) pending_async: HashSet<Id>,
+}
+
+struct MyWaker {
+    id_path: IdPath,
+    req_chan: SyncSender<IdPath>,
+}
+
+impl ArcWake for MyWaker {
+    fn wake_by_ref(arc_self: &Arc<Self>) {
+        //println!("path = {:?}", arc_self.id_path);
+        let _ = arc_self.req_chan.send(arc_self.id_path.clone());
+    }
+}
+
+impl Cx {
+    pub(crate) fn new(req_chan: &SyncSender<IdPath>) -> Self {
+        Cx {
+            id_path: Vec::new(),
+            req_chan: req_chan.clone(),
+            pending_async: HashSet::new(),
+        }
+    }
+
+    pub fn push(&mut self, id: Id) {
+        self.id_path.push(id);
+    }
+
+    pub fn pop(&mut self) {
+        self.id_path.pop();
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.id_path.is_empty()
+    }
+
+    pub fn id_path(&self) -> &IdPath {
+        &self.id_path
+    }
+
+    /// Run some logic with an id added to the id path.
+    ///
+    /// This is an ergonomic helper that ensures proper nesting of the id path.
+    pub fn with_id<T, F: FnOnce(&mut Cx) -> T>(&mut self, id: Id, f: F) -> T {
+        self.push(id);
+        let result = f(self);
+        self.pop();
+        result
+    }
+
+    /// Allocate a new id and run logic with the new id added to the id path.
+    ///
+    /// Also an ergonomic helper.
+    pub fn with_new_id<T, F: FnOnce(&mut Cx) -> T>(&mut self, f: F) -> (Id, T) {
+        let id = Id::next();
+        self.push(id);
+        let result = f(self);
+        self.pop();
+        (id, result)
+    }
+
+    pub fn waker(&self) -> Waker {
+        futures_task::waker(Arc::new(MyWaker {
+            id_path: self.id_path.clone(),
+            req_chan: self.req_chan.clone(),
+        }))
+    }
+
+    /// Add an id for a pending async future.
+    ///
+    /// Rendering may be delayed when there are pending async futures, to avoid
+    /// flashing, and continues when all futures complete, or a timeout, whichever
+    /// is first.
+    pub fn add_pending_async(&mut self, id: Id) {
+        self.pending_async.insert(id);
+    }
+}
