+++ src/view/mod.rs
@@ -24,12 +24,14 @@ mod button;
 // mod use_state;
 mod linear_layout;
 mod list;
-mod sequence;
 mod view;
+mod typed_view;
+mod cx;
 
-pub use any_view::AnyView;
+pub use any_view::AnySequence;
 pub use button::button;
 pub use linear_layout::{h_stack, v_stack, LinearLayout};
 pub use list::{list, List};
-pub use sequence::ViewSequence;
-pub use view::{Cx, View, ViewMarker};
+pub use view::{View, ViewSequence};
+pub use typed_view::TypedView;
+pub use cx::Cx;
